refactor renderer to match RFC
Main change was to get it to match the Renderer interface by setting all
the supporting parameters (path, etc.) on the struct so you can only
pass the contents to the Render call.

Also changed the Backup boolean to a BackupFunc function that takes the
path to the file to backup as its argument. Included the previous backup
functionality as a public function that can be used with it.