rework Watcher with cache and Wait
The data cache (used to be brains) lives on the watcher as it is
responsible for keeping it up to date.

Wait() method is added along with tests. It replaces where the old code
published the channels, instead encapsulating the functionality they
provided.. blocking until new data (or error) was received.