refactor Set data struct and use
The existing Set implementation under dependency/ was not used anywhere
in dependency/ so it made no sense to keep it there. I started by moving
it up to the top level.

I then had the need for a simple string based set and had a simple one
already made. I determined I could combine them and get one very simple
string based set and an additional one, that embedded the first, that
would provide the *view storage needed by the existing code.

This removes the old set, adds the new one and tweaks things to use the
new version.