Rework public API for Resolver logic
While working on the Resolver (the code that evaluates the template and
update the dependency tracking) it became obvious that the Template and
Watcher passed in needed to be interfaces for testing. While working on
that I came up with these changes for the APIs.

New APIs on Watcher:
    Changed(tmpl-id)bool - returns true if the template data has changed
    TemplateDeps(tmpl-id, ...dependency) - tell watcher what
        dependencies the template has (used by Changed)

Made Public: (needed to be public to be in the public interface)
on Watcher:
    Add(dependency)bool - adds a new dependency
on Template:
    ID()string - need to be able to reference template ids for above

Also made watching() private (was Watching). Might need to change it
back once I start work on the consul-template use case but want to try
it being private first as I felt the API was getting a bit to large.