make Dependency interface publicly available
It is used with public calls, so needs to be public. It is still used in
internal/dependency/, but only to check that all the objects implement
it correctly. So the main definition still lives in the internal
location but it is made public with a simple sub-type.

We could move it up, but then there'd be 2 versions of it unless we
eliminated the type check in the internal code which is kind of nice.
This seems the best trade off for now.