factor out http.transport creation and config
The first step in allowing it to be replaced for testing.

Eliminated a bunch of duplicate code. Part of that was eliminating the
separate vault/consul inputs structs as this let me further reduce the
amount of code and simplified refactoring out the transport code. They
were identical excpet for a couple fields which will just be unused on
the other (ie. they don't conflict).