set HttpClient field on both API clients
Previosly they were setting the transport for consul and the http client
on vault. I checked and consul also supported the http client so I
simplified things by having both use it.

This also changes the client_set creation input parameter to accept an
http client. Having the http client as an input parameter allows it to
be adjusted for testing.