tweak FuncMapMerge processing/func signatures
Change the function signature for dependency functions to have the
return type be 'interface{}'. The FuncMap's values are interface{}
already, so it matches and works with FuncMap fine and it makes matching
and calling them work no matter the returned function signature (which
can vary quite a bit).

Also changed the documentation to specify the new expected signature.

This is all about being able to assign methods to the FuncMap that
require the recaller, missing and used variables for proper dependency
tracking.