add a fake dependency that supports slice data
Added FakeListDep that returns a slice for Fetch allowing it to be used
to test multi-pass dependencies (slice/list entries are used in loop
with another dependency).

Changed FakeDep to return its Name field as the data instead of a
hard coded string which makes it easier to write tests as you don't need
to remember the string to check for in the evaluated template.

Also tweaked the others to only sleep for a Microsecond (was a
Millisecond which seemed overly long).