make test channels buffered
The real channels used are buffered, so this better reflects what is
being tested.

Without the buffer the firing of the select{} in the test was less
determinant and sometimes the close of the done channel (which happens
at the end of the function being tested) would fire instead of the
success channel and making the success channel buffered fixes that.