fix issue with empty content not rendering to file
Problem with initial run when template resolves to the content being an
empty string. This is compared vs the existing file contents which
defaults to an empty string if the file doesn't exist. Thus if the file
doesn't exist and the template resolves to an empty string the file is
not written (even though it should be).

Also add a bunch of renderer.Render tests, including one for this.

Originally filed as an issue on consul-template.
Consul-template issue: github.com/consul-template/issues/1393