remove internal data structure from interface
The Dependency interface needs to be exported outside hashicat as it is
used in several of the public APIs. Defined in internal/dependency/ it
is internal only and while we can bring it up to the top and export it,
we cannot do that with the data structures it references without
changing their type or using type aliasing. Type aliasing internal
structures in this way is bad as it breaks documentation and IDE/gopls
meta data.

So this change pulls the QueryOptions argument used in the Fetch()
method into a separate setOptions() method that is defined on a second
interface which is only used internally. This lets us keep the options
setting support while cleaning it out of the public interface.