indicate blocking dependency by type
Blocking dependency queries were indicated as such by returning a flag
in the ResponseMetadata returned by Fetch(). Whether the query is
blocking or not is specific to the way the query is made and has nothing
to do with the returned data. This change converts the way the
information about blocking is conveyed to typing data instead of
returned meta-data.

This was done as part of work to eliminate the structures from the
external/public Dependency interface which includes the ResponseMetadata
struct which is (in the process of) being eliminated.

The main addition is the BlockingQuery interface that the queries that
block now meet so you can assert that type to check.