move internal but public dep types into own pkg
This starts a public dependency sub-package that will contain the
required types to create dependencies outside the library. It is a place
where we can put more public parts as they come up.

For now it only contains the type information to allow public and
internal use without exporting aliases or having circular references.