calling Stop now triggers Wait to return
Wait can block forever so Stop needs to force it to return so we don't
leak goroutines. This adds that as well as tests.

Synchronization was tricky in the tests are required another
notification channel as well as the internal stop channel. This
complexity could go away if we decided that a watcher is no-good after
you call Stop and if you still need it, you'd make a new one.