propagate errors from client setup to caller
The error handling in the client setup was only half done and didn't
properly propagate errors up the stack. This fixes that and tweaks the
test to check that and work w/o starting consul.

Includes a slight tweak to Consul/Vault functions so they don't panic if
called before being setup.