remove child sub-module
This was pulled in as it was a nicely self-contained sub-module of
consul-template, so I thought it logical to include in a library
based on consul-template. However the focus of the library is on
configuration and templating... not child process management. None of
the existing prjects this library is tagetting (sync, nomad, vault) use
the child functionlity and it seems out of place.

Removing it and it will remain as part of consul-template for now.