remove non-dependency related template functions
Remove all the utility/helper template functions that were in
consul-template but aren't the focus of hashicat. With hashicat we want
to focus on the external dependencies/data, monitoring and triggering
re-renders. Keeping the helper functions needed to extend that for a
specific application to that application.

This strips out all the functions and their tests.