strip out rest of non-dependency functions
As we're moving to having a standard library of these non-dependency
using functions, it makes sense to also move these there.

This removes the consul filtering functions (byMeta, etc.) as well as
the env and file functions. They will be added back with the library
along with most of the rest.