remove old/redundant data received map
This map was added back (in consul-template) when they used a different
internal map for each data type and it was expensive and ugly to check
them all every time, so they introduced this field to simplify checking.
Later they converted to a single map[string]interface{} to hold all the
data but didn't remove the (now redundant) map that tracked received.