large refactor of dependency tracking/management
Reworks how it keeps track of dependencies/views, templates and their
relationships like which dependencies are used by which templates and
visa versa.

Main reasons for this.
1. To many objects with overlapping responsibilities, this cuts things
   down from ~6 objects to 1 (including sub-structs, ~10 to 3).
2. It eliminates the shared tracking of changes and localizes that to
   allow for per-template level checks.
3. Notifies all templates watching a dependency when that dep has
   updated.
4. Reduces in-memory redundant caches.

2 above will allow for per-template async management

Also includes some simplifications to the related interfaces as some
methods were changed, added or made obsolete.
The resulting interfaces are smaller as well.