Up timeout on waiting on data chan to 1 second
up from 1 microsecond

The original intent of that is to stall just a tiny bit in case a bunch
of data is coming down the pipe and we just want to make sure it can get
them.

We currently have a bug where there is a race between the incoming data
and the template rendering when there are a lot of dependencies coming
in for the template on the first run.

Raising it to 1 second should give more updates a chance to come in
before it moves on and renders the template. Mitigating that initial,
first run flood.

This is a temporary workaround and requires a revisit of how data comes
in and how that is tracked.