track first use of data for completeness checks
This adds a data point to tracked dependencies that marks them as having
been used at least once. It is set when the data is first pulled from
the cache by the Recaller.

This is then used to determine the Complete-ness of the template
(notifier), requiring that all data has been retrieved and used. This
fixes the race between the template rendering (using the fields) and the
completeness calculation as it cannot be counted as complete anymore
just due to the data having been fetched (but not used).