explicit registering of notifiers with watcher
- adds a 'name' field to templates that gets appended to the ID to ensure
  they are unique and can be added the the watcher
- 'name' field is optional, ID defaults to just the md5hex if not set
- explicit Register call to the watcher with checks to ensure each Notifier
  is unique and returns an error if registering a duplicate
- attempting to use unregistered Notifier results in a panic
- changes the old Register/register methods to Track/track which more
  closely aligns with their use (and freed up Register)

These semantics should allow multiple parallel uses of the same template
content, with different instances to work while letting the developer
structure the code to ensure everything lines up.

It eliminates the "auto-registering on first use" that was used before
as there was no way for it to properly check (and provide an error) that
the Notifier was a duplicate.