fix buffered to work with new registration setup
The Buffer() call was testing that the template/notifier had called by
checking that it had been registered. This worked with the old system of
registering on first use, but with pre-registration required this no
longer works.

This change makes it check that the notifier has been referenced in
relation to a dependency (a tracked_pair w/ that notifier in it) as this
means the template has been called at least once (which is what we want).