Skip retrying fetch on 400
When querying Consul OSS with the namespace query param, it results
in a 400. And when retry is configured, it attempts to retry a bad
request. Changes now return on first 400 instead of waiting through
all the retries.

on the client side, it is preceived as hanging and blocks waiting
on expected behavior.