temp fix to allow blocking external dependencies
Blocking dependenceis set their blocking options using the
QueryOptionsSetter interface, but this interface takes QueryOptions
which is/was not public. So it was impossible to implement the
QueryOptionsSetter interface.

This uses type aliases to raise them out of the internal/dependency
module and make them public so it is possible to implement the interface.

This is a temporary fix to address things until we are able to refactor
the internal packages to be more modular.