update dependencies to have ID method
Dependencies had a legacy base of using String() for object IDs. With
the various refactorings IDs have become more important and are used
across object types. Got tired of dependencies being out of sync witht
he rest of the code on this so went through and added ID to all
dependencies and made their String calls use it (for convenience in
debugging and places where the Stringer interface will be used).

Should be absolutely no behavioral changes. Just the required changes to
add ID and use it.