vault default lease parameter w/ no default
Remove the hardcoded default lease duration instead always using the
value passed to the watcher. Previously it supported replacing the lease
parameter with a hardcoded default (5 min) if the paremeter was a 0.

So this change means that sort of logic needs to be handled in the
application and the library only works with what's provided.

An exception is the hardcoded vault token dependency as it only used the
hardcoded value. It uses the hardcoded value locally now as the module
level one has been removed.

Ported* over from consul-template.
PR: #1446

*These changes are not a direct port from consul-template as it just
uses the module default as a variable and sets it as a global option.
This changes things so consul-template can just pass the default lease
duration config setting to the watcher (and implement a hardcoded
default before that if needed).