fixes vault KvV2 list and write path issue
Vault KvV2 list, `secrets` template function, didn't deal with the KvV2
path changes so they had to be manually set. Same with the write
functionlaity, `secret` used as write w/ 2 args. This change makes sure
the paths are correct behind the scenes so dealing with the path changes
for the KV version isn't left to the end user.

Ported from consul-template.
PR: #1468