tie buffer period triggers to data updates
Fixes a bug with endless buffer triggers if you call resolver.Run with
a template that hasn't updated.

Ties the buffer period to only be triggered by data updates instead of
being tied to template resolver runs.

Buffering is now always on once enabled (used to be skipped on first run).