fixed test
Test of watcher.Stop during watcher.Watch as erroneously passing when
context timed out instead of stopCh returning. It wasn't testing stopCh
as that gets cleared out at the start of watcher.Watch and there was a
race between that and the test function calling Stop. That is it would
call Stop before calling watcher.Watch.