update Collector interface with new method name
Mark -> MarkForSweep

Changed this on the watcher and forgot to update the interface. Missed
it as the only check was the dynamic check that allows for it to fail on
purpose.

Moves the Collector interface into the resolver where it is used.

Updates the Collector interface and use in Resolver with the new method
name.

Adds a Collector type assertion in the watcher.go file to cause a
compiler error if I do this again in the future.