revert not storing nil/zero values in cache
I changed the logic of the cache to only save real values, not caching
any invalid, nil, or zero values (tested using reflect module). This
broke things in CTS as it changed the behavior of Recall() to return
false for the second, boolean value as the nil value wasn't stored in
the cache map. So the nil value was the same, but the bool was different
and this broke update detection in place.