make template.Notify()'s bool return meaningful
Change template.Notify() to return true if it marks the template as
dirty, returns false if it has already been marked as dirty.

This change is to template clarify's the meaning of the Notify's boolean
return value as well as making it play nicer with the watcher to help
eliminate duplicate template notifications.