Revert "make template.Notify()'s bool return meaningful"
This reverts commit 97c05d16bb8c8d6ff6d69c12dfbf8e691bcce9a1.

Reverting this as the new semantics broke CTS and while it does increase
the amount of information available when viewed in relation to it's
current use in watcher it mainly changes the semantics of when the
watcher returns data about the template being ready. From (with the
reverted change) a template being ready for rendering being only sent
when it becomes ready to (with this reverted) to it sending every time
checked when it's ready.