eliminate temporary goroutine leaking due to sleep
Sleep() calls during dependency Fetch and similar actions that take
place in goroutines cause them to leak for the duration of the sleep and
possibly cause delays in exiting or uncontrolled memory growth if they
happen fast enough.

This replaces those cases of Sleep() with select cases on <-time.After()
(for the sleep) and <-stopCh (to prevent leak).